+++ doc/changelog.rst
@@ -3,6 +3,7 @@
 Release Notes
 =============
 
+- :feature:`656` Instead of by average, review scores are now shown aggregated by the median.
 - :bug:`655` Mails to submissions with titles near the length cap (of 200 characters) could not be created, since with the added "[event]" prefix the mail subject was too long to be saved.
 - :feature:`-` Question pages now feature visual statistics where appropriate.
 - :feature:`-` There now is a page for submission statistics.